requirements.txt
Flask 
Flask-Login 
Flask-Bcrypt 
Flask-WTF 
FLask-Migrate 
Flask-SQLAlchemy 
pyotp 
qrcode 
python-decouple
Pillow
pytest
beautifulsoup4

